pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "hobbes"
version = "0.1.0"
description = "A package manager for GitHub release binaries"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "Your Name", email = "your@email.com" }
]
keywords = ["package-manager", "github", "binary", "cli"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Build Tools",
]
dependencies = [
    "click>=8.0",
    "httpx>=0.25",
    "pyyaml>=6.0",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-httpx>=0.21",
]

[project.scripts]
hobbes = "hobbes.cli:main"

[project.urls]
Homepage = "https://github.com/yourusername/hobbes"
Repository = "https://github.com/yourusername/hobbes"

[tool.hatch.build.targets.wheel]
packages = ["src/hobbes"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
